Changes to msg receiver syncing and socket usage.
The message receiver does no longer automatically run in it's own
thread. NetworkManager used to automatically spawn a thread to run
it, but now this is done explicitly through the new method
NetworkManager::runMessageReceiverInThread(). So, the client code
can choose to call NetworkManager::runMessageReceiverInThread(),
NetworkManager::runMessageReceiver() (blocking until connection
closes) or call NetworkManager::runMessageReceiver() from a thread
managed by the client.

The TCPConnection class has a new addition as well;
bool32 TCPConnection::waitForSocket(uint32 timeoutSeconds,
                                    uint32 timeoutMicroseconds)
It uses select to wait for the socket to be ready for reading/
calling recv(). This was needed in runMessageReceiver() for it
to be able to receive the exit signal from the destructor without
having to wait for the server to send something first.