class MsgPrefix and frontend stdin multiplexing.
Implemented class MsgPrefix, which represents the prefix of a IRC
message, if any. If the prefix can be parsed into nick!user@host
parts, that will be mirrored in the MsgPrefix instance, and the
various components can be easily accessed. PRIVMSG and TOPIC now
uses the MsgPrefix as origin argument.

frontend_console_cpp has been given the ability to read user input
from stdin, without blocking indefinetely. This input is sent
straight to the irc-server, unless it is the "die" command, which
kills the connection and in the end, the program.